chore: moved files from adapters into their respective folders
